+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0
+
+- Add zustand store
+- Add autocomplete to find participants or add new ones
+
 # 0.1.2
 
 ### FEATURES